# api/openapi/schemas.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Steel Browser Instance API",
    "description": "Documentation for controlling a single instance of Steel Browser",
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {},
    "schemas": {
...
      "ScreenshotRequest": {
        "title": "ScreenshotRequest",
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "proxyUrl": {
            "type": "string",
            "nullable": true,
            "description": "Proxy URL to use for the scrape. Provide `null` to disable proxy. If not provided, current session proxy settings will be used."
          },
          "delay": {
            "type": "number"
          },
          "fullPage": {
            "type": "boolean"
          },
          "logUrl": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ScreenshotResponse": {
        "title": "ScreenshotResponse"
      },
...
    }
  },
  "paths": {
...
    "/v1/screenshot": {
      "post": {
        "operationId": "screenshot",
        "summary": "Take a screenshot",
        "tags": [
          "Browser Actions"
        ],
        "description": "Take a screenshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenshotResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://0.0.0.0:3000/",
      "description": "Local server"
    },
    {
      "url": "http://0.0.0.0:3000",
      "description": "Local server from env variables"
    }
  ]
}